91896 ordering_system v2
Ordering system for a smoothie bar.
System takes inputs and outputs it in the orders menu. Users have a choice in milk types and smoothie flavours. Customer orders are stored, and there is a function to delete orders. There are no unacceptable inputs.

Expected:
System accepts names from the user that are names that would be used in the modern world, such as ‘Alice’, ‘Lucy’, ‘Katherine’.
System does not crash if a strange input is given. If you provide five paragraphs of Lorem Ipsum as the name, the system will take that as the customer name and output it. If you put a space (` `) as the name, the system will take that as the customer name and output it. What you provide as the name, the system will take that as the customer name, then output it.
If there is no input, system will continue and ignore the missing input. If you do not provide a name, the system will output the name as blank. If you do not select a milk, the system will output the milk as blank. If you do not select any smoothie flavours, the system will output the items as blank, and the total cost as $0.
System works for all inputs. Not matter what is entered, the system will continue.
Boundary:
Because all inputs are accepted, there are no boundary inputs.
Invalid:
Because all inputs are accepted, there are no invalid inputs.