91896 ordering_system v3
Ordering system for a smoothie bar.
System takes inputs and outputs it in the orders menu. Users have a choice in milk types and smoothie flavours. Customer orders are stored, and there is a function to delete orders.

Expected:
System accepts names from the user that are names that would be used in the modern world, such as ‘Alice’, ‘Lucy’, ‘Katherine’.
System does not crash if a strange input is given. If you provide five paragraphs of Lorem Ipsum as the name, the system will take that as the customer name and output it. If you put a space (` `) as the name, the system will take that as the customer name and output it. What you provide as the name, the system will take that as the customer name, then output it as an alert.
If there is no input, system will stop and ask for the missing input. If any inputs are missing, the order will not be pushed through to the order list.
Boundary:
Boundary inputs are if only one character is entered into the customer name input. The system accepts these inputs as valid, and allows the system to continue. If you put a W as the name, the system will take that as the customer name and output it. If you put a P as the name, the system will take that as the customer name and output it.
Invalid:
Invalid inputs are if not all inputs are provided. If you do not provide a name, the system will say that you have not provided a name. If you do not select a milk, the system will say that you have not provided a milk type. If you do not select any smoothie flavours, the system will say that you have not selected any smoothie flavours.