91896 ordering_system v6
Ordering system for a smoothie bar.
System takes inputs and outputs it in the orders menu. Users have a choice in milk types and smoothie flavours. Customer orders are stored, and there is a function to delete orders.
The smoothie flavour list is appended from a JavaScript constant to make the ordering system more flexible and robust. The multiple ‘if’ statements have been compiled into one selection utilising variables, to make it easier to add smoothie flavours to the order options. The flavours and their prices have been separated to further increase the ease of use.

Expected:
System accepts names from the user that are names that would be used in the modern world, such as ‘Alice’, ‘Lucy’, ‘Katherine’.
System does not crash if a strange input is given. If you provide five paragraphs of Lorem Ipsum as the name, the system will take that as the customer name and output it. If you put a space (` `) as the name, the system will take that as the customer name and output it. What you provide as the name, the system will take that as the customer name, then output it as an alert.
If there is no input, system will stop and ask for the missing input. If any inputs are missing, the order will not be pushed through to the order list.
If the number provided to remove an order is more than 0, the system accepts the order number, removes the specified order, and updates the list. If that order number does not have an order associated with it, nothing happens. If you provide 1 as the order number, the system removes the second order, and updates the list. If you provide 2 as the order number, the system removes the third order, and updates the list.

Boundary:
Boundary inputs are if only one character is entered into the customer name input. The system accepts these inputs as valid, and allows the system to continue. If you put a W as the name, the system will take that as the customer name and output it. If you put a P as the name, the system will take that as the customer name and output it.
Boundary inputs are also if 34 characters are entered into the customer name input. The system accepts these inputs as valid, and allows the system to continue. If you put ‘qwertyuiopasdfghjklzxcvbnm12345678’ as the name, the system will take that as the customer name and output it.
Boundary inputs are if the number to remove an order is 0. The system accepts 0 as an order number, removes the first order, and updates the list.

Invalid:
Invalid inputs are if not all inputs are provided. If you do not provide a name, the system will say that you have not provided a proper name. If you do not select a milk, the system will say that you have not provided a milk type. If you do not select any smoothie flavours, the system will say that you have not selected any smoothie flavours.
Invalid inputs are also if 35 characters are entered into the customer name input. The system does not accept these inputs as valid. If you put ‘qwertyuiopasdfghjklzxcvbnm123456789’ as the name, the system will say that you have not provided a proper name.
If the number provided to remove an order is less than 0, the system alerts that the order number to remove is too small. If you provide -2 as the order number, the system does not accept it as a valid order number to remove.